--- a/src/compas_view2/objects/bufferobject.py
+++ b/src/compas_view2/objects/bufferobject.py
@@ -37,6 +37,16 @@ class BufferObject(Object):
         self.pointsize = pointsize
         self.opacity = opacity
         self.background = False
+        self._bounding_box = None
+        self._bounding_box_center = None
+
+    @property
+    def bounding_box(self):
+        return self._bounding_box
+
+    @property
+    def bounding_box_center(self):
+        return self._bounding_box_center
 
     def make_buffer_from_data(self, data):
         """Create buffers from point/line/face data.
@@ -87,7 +97,9 @@ class BufferObject(Object):
     def make_buffers(self):
         """Create all buffers from object's data"""
         if hasattr(self, '_points_data'):
-            self._points_buffer = self.make_buffer_from_data(self._points_data())
+            data = self._points_data()
+            self._points_buffer = self.make_buffer_from_data(data)
+            self._update_bounding_box(data[0])
         if hasattr(self, '_lines_data'):
             self._lines_buffer = self.make_buffer_from_data(self._lines_data())
         if hasattr(self, '_frontfaces_data'):
@@ -114,6 +126,12 @@ class BufferObject(Object):
         """Update the object"""
         self._update_matrix()
         self.update_buffers()
+
+    def _update_bounding_box(self, positions=None):
+        """Update the bounding box of the object"""
+        positions = np.array(positions or self._points_data()[0])
+        self._bounding_box = np.array([positions.min(axis=0), positions.max(axis=0)])
+        self._bounding_box_center = np.average(self.bounding_box, axis=0)
 
     def draw(self, shader, wireframe=False, is_lighted=False):
         """Draw the object from its buffers"""