@brgcode 2021-03-16
how does this scale to a large number of objects?