@Licini 2021-03-17
> how does this scale to a large number of objects?

It does not make difference when there is no transparent objects. Some benchmark:

1000 opaque vs transparent cubes:   8 fps -> 6 fps
500 opaque vs transparent cubes:   15 fps -> 12 fps